[KAN-52] Fix: 카프카 빈 설정 및 docker-compose 에러 해결
