+++ .gitignore
@@ -43,3 +43,6 @@ src/main/resources/application.properties
 
 ### smart tomcat ###
 .smarttomcat
+
+### Environment ###
+.env
