+++ docker-compose.yml
@@ -28,7 +28,9 @@ services:
       - "9092:9092"
     environment:
       KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
+      KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9092
       KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka:9092
+      KAFKA_BROKER_ID: 1
     depends_on:
       - zookeeper
     container_name: kafka
@@ -45,17 +47,23 @@ services:
       - "1935:1935"
       - "8081:8080"
     volumes:
-      - ./nginx.conf:/etc/nginx/nginx.conf:ro
+      - ./nginx.conf:/etc/nginx/nginx.conf
     container_name: nginx-rtmp
 
   themoonha-backend:
     build:
       context: .
     ports:
-      - "8080:8080"
+      - "8082:8080"
     depends_on:
       - kafka
       - nginx-rtmp
     environment:
       SPRING_PROFILES_ACTIVE: "docker"
+      TNS_ADMIN: /app/config/wallet
+      S3_ACCESS_KEY: ${S3_ACCESS_KEY}
+      S3_SECRET_KEY: ${S3_SECRET_KEY}
+      S3_BUCKET_NAME: ${S3_BUCKET_NAME}
+    volumes:
+      - /Users/jingyu/TheMoonHa/Wallet_THEMOONHA:/app/config/wallet
     container_name: themoonha-backend
