--- src/main/java/com/innerpeace/themoonha/global/config/WebSocketConfig.java
@@ -1,40 +1,0 @@
-package com.innerpeace.themoonha.global.config;
-
-import com.innerpeace.themoonha.global.handler.StreamingWebSocketHandler;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.*;
-
-/**
- * 웹 소켓 설정
- *
- * @author 김진규
- * @since 2024.09.01
- * @version 1.0
- *
- * <pre>
- * 수정일        수정자        수정내용
- * ----------  --------    ---------------------------
- * 2024.09.01  김진규        최초 생성
- * </pre>
- * @since 2024.09.01
- */
-@Configuration
-@EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketConfigurer, WebSocketMessageBrokerConfigurer {
-    @Override
-    public void registerWebSocketHandlers(WebSocketHandlerRegistry webSocketHandlerRegistry) {
-        webSocketHandlerRegistry.addHandler(new StreamingWebSocketHandler(), "/streaming").setAllowedOrigins("*");
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        registry.enableSimpleBroker("/topic");
-        registry.setApplicationDestinationPrefixes("/app");
-    }
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/ws").setAllowedOrigins("*").withSockJS();
-    }
-}
