+++ src/main/webapp/WEB-INF/spring/security-context.xml
@@ -22,7 +22,7 @@
     <http pattern="/lesson/**" security="none"></http>
     <http pattern="/bite/**" security="none"></http>
     <http pattern="/craft/**" security="none"></http>
-    <http pattern="/live/**" security="nono"></http>
+    <http pattern="/live/**" security="none"></http>
 
     <!-- HTTP 보안 설정 -->
     <http auto-config="true" use-expressions="true">
